+++ libraries/decoder_opus/src/main/java/androidx/media3/decoder/opus/LibopusAudioRenderer.java
@@ -29,6 +29,7 @@ import androidx.media3.exoplayer.audio.AudioRendererEventListener;
 import androidx.media3.exoplayer.audio.AudioSink;
 import androidx.media3.exoplayer.audio.AudioSink.SinkFormatSupport;
 import androidx.media3.exoplayer.audio.DecoderAudioRenderer;
+import androidx.media3.extractor.VorbisUtil;
 
 /** Decodes and renders audio using the native Opus decoder. */
 @UnstableApi
@@ -138,6 +139,12 @@ public class LibopusAudioRenderer extends DecoderAudioRenderer<OpusDecoder> {
     @C.PcmEncoding
     int pcmEncoding = decoder.outputFloat ? C.ENCODING_PCM_FLOAT : C.ENCODING_PCM_16BIT;
     return Util.getPcmFormat(pcmEncoding, decoder.channelCount, OpusDecoder.SAMPLE_RATE);
+  }
+
+  @Nullable
+  @Override
+  protected int[] getChannelMapping(OpusDecoder decoder) {
+    return VorbisUtil.getVorbisToAndroidChannelLayoutMapping(decoder.channelCount);
   }
 
   /**